Moves the donation link to the CC pages
The donation pages maintained for CC are superior to the ones that P1T2
was using. This points them to those pages to allow for a single page to
maintain.

There are also updatees to the included states to reflect the changes as
well as moving the national page to the same CC pages